pyproject.toml
[tool.poetry]
name = "talk2everything-app"
version = "0.1.0"
description = "Talk2Everything – FastAPI API and agent for AI chat"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
python-dotenv = "^1.0.0"
langchain = "*"
langchain-oci = "*"
langgraph = "*"

oci-openai = "*"
mcp = "*"
fastmcp = "^2.0.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
pydantic-settings = "^2.0.0"
pypdf = "^4.0.0"
oracledb = "*"
pandas = "^2.0.0"
numpy = "^1.24.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
httpx = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
